% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_calificacion.R
\name{get_calificacion}
\alias{get_calificacion}
\title{Obtener calificación Renacyt}
\usage{
get_calificacion(
  puntaje_formacion = 0,
  puntaje_produccion = 0,
  puntaje_asesoria = 0,
  indice_h = FALSE
)
}
\arguments{
\item{puntaje_formacion}{Puntaje obtenido en la sección Formación}

\item{puntaje_produccion}{Puntaje obtenido en la sección Producción}

\item{puntaje_asesoria}{Puntaje obtenido en la sección Asesoría}

\item{indice_h}{logical. ¿El investigador tiene Índice H mayor a 10?}
}
\value{
character. Estado de calificación y nivel correspondiente.
}
\description{
Permite calcular si el investigador pasa la valla mínima en
distintos escenarios y el nivel correspondiente a su calificación RENACYT.
}
\examples{
get_calificacion(2, 10)
get_calificacion(1, 6)
get_calificacion(1, 200, 0, TRUE)
}
