% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_puntaje_asesoria.R
\name{get_puntaje_asesoria}
\alias{get_puntaje_asesoria}
\title{Calcular puntaje en Asesoria}
\usage{
get_puntaje_asesoria(n_doct = 0, n_mag = 0, n_bach = 0)
}
\arguments{
\item{n_doct}{Cantidad de tesis de doctorados asesoradas}

\item{n_mag}{Cantidad de tesis de maestria asesoradas}

\item{n_bach}{Cantidad de tesis de bachiller asesoradas}
}
\value{
Un numero real
}
\description{
Permite calcular los puntos obtenidos por el investigador en asesorías
de tesis en distintos niveles de estudios
}
\examples{
get_puntaje_asesoria(n_doct = 2)
get_puntaje_asesoria(n_mag = 10)
get_puntaje_asesoria(n_mag = 10, n_bach = 7)
}
