% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_puntaje_formacion.R
\name{get_puntaje_formacion}
\alias{get_puntaje_formacion}
\title{Calcular puntaje en Formacion}
\usage{
get_puntaje_formacion(grado_obtenido = "Ninguno")
}
\arguments{
\item{grado_obtenido}{Grado academico obtenido. Debe estar registrado en SUNEDU.

El grado obtenido debe pertenecer al listado c("Doctor", "Magister", "Título profesional",
"Bachiller o egresado", "Constancia de matrícula (Estudiante)", "Ninguno")}
}
\value{
Un numero entero entre 0 y 10 cerrado
}
\description{
Permite  calcular los puntos obtenidos por el investigador
referidos a su formacion profesional
}
\examples{
get_puntaje_formacion("Doctor")
}
