% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_puntaje_produccion.R
\name{get_puntaje_produccion}
\alias{get_puntaje_produccion}
\title{Calcular puntaje en Producción}
\usage{
get_puntaje_produccion(
  q1 = 0,
  q2 = 0,
  q3 = 0,
  q4 = 0,
  conf = 0,
  pat_invencion = 0,
  pat_modelo = 0,
  n_libros = 0,
  n_cap = 0
)
}
\arguments{
\item{q1, q2, q3, q4}{Cantidad de artículos publicados en el cuartil de citado correspondiente}

\item{conf}{Cantidad de conference proceedings indexadas}

\item{pat_invencion}{Cantidad de patentes de invención}

\item{pat_modelo}{Cantidad de patentes de modelo}

\item{n_libros}{Cantidad de libros publicados}

\item{n_cap}{Cantidad de capítulos de libro publicados}
}
\value{
Un numero entero
}
\description{
Permite  calcular los puntos obtenidos por el investigador
referidos a su producción de artículos, patentes y libros.
}
\examples{
get_puntaje_produccion(q1 = 2, conf = 1, n_cap = 2)
}
