% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_puntaje_total.R
\name{get_puntaje_total}
\alias{get_puntaje_total}
\title{Obtener puntaje total}
\usage{
get_puntaje_total(
  puntaje_formacion = 0,
  puntaje_produccion = 0,
  puntaje_asesoria = 0
)
}
\arguments{
\item{puntaje_formacion}{Puntaje obtenido en la sección Formación}

\item{puntaje_produccion}{Puntaje obtenido en la sección Producción}

\item{puntaje_asesoria}{Puntaje obtenido en la sección Asesoría}
}
\value{
Un número real
}
\description{
Obtiene el puntaje total a partir de los datos del investigador.
En realidad, solo suma los puntajes de los tres módulos.
}
\examples{
get_puntaje_total(10, 5, 10)
}
